Split some WMS tests
