+++ tests/src/python/CMakeLists.txt
@@ -233,6 +233,9 @@ IF (WITH_SERVER)
   ADD_PYTHON_TEST(PyQgsServer test_qgsserver.py)
   ADD_PYTHON_TEST(PyQgsServerPlugins test_qgsserver_plugins.py)
   ADD_PYTHON_TEST(PyQgsServerWMS test_qgsserver_wms.py)
+  ADD_PYTHON_TEST(PyQgsServerWMSGetMap test_qgsserver_wms_getmap.py)
+  ADD_PYTHON_TEST(PyQgsServerWMSGetLegendGraphic test_qgsserver_wms_getlegendgraphic.py)
+  ADD_PYTHON_TEST(PyQgsServerWMSGetPrint test_qgsserver_wms_getprint.py)
   ADD_PYTHON_TEST(PyQgsServerSettings test_qgsserver_settings.py)
   ADD_PYTHON_TEST(PyQgsServerProjectUtils test_qgsserver_projectutils.py)
   ADD_PYTHON_TEST(PyQgsServerSecurity test_qgsserver_security.py)
